Use generics in production app
