+++ examples/production-app/graphql/Connection.ts
@@ -1,0 +1,27 @@
+/** @gqlType */
+export type Connection<T> = {
+  /** @gqlField */
+  edges: Edge<T>[];
+  /** @gqlField */
+  pageInfo: PageInfo;
+};
+
+/** @gqlType */
+export type Edge<T> = {
+  /** @gqlField */
+  node: T;
+  /** @gqlField */
+  cursor: string;
+};
+
+/** @gqlType */
+export type PageInfo = {
+  /** @gqlField */
+  startCursor: string | null;
+  /** @gqlField */
+  endCursor: string | null;
+  /** @gqlField */
+  hasNextPage: boolean;
+  /** @gqlField */
+  hasPreviousPage: boolean;
+};
