--- examples/production-app/models/PageInfo.ts
@@ -1,11 +1,0 @@
-/** @gqlType */
-export type PageInfo = {
-  /** @gqlField */
-  startCursor: string | null;
-  /** @gqlField */
-  endCursor: string | null;
-  /** @gqlField */
-  hasNextPage: boolean;
-  /** @gqlField */
-  hasPreviousPage: boolean;
-};
