+++ examples/production-app/models/PostConnection.ts
@@ -3,31 +3,15 @@ import * as DB from "../Database";
 import { Ctx } from "../ViewerContext";
 import { Query } from "../graphql/Roots";
 import { Post } from "./Post";
-import { PageInfo } from "./PageInfo";
+import { Connection } from "../graphql/Connection";
 import { connectionFromArray } from "graphql-relay";
-
-/** @gqlType */
-export type PostConnection = {
-  /** @gqlField */
-  edges: PostEdge[];
-  /** @gqlField */
-  pageInfo: PageInfo;
-};
 
 /**
  * Convenience field to get the nodes from a connection.
  * @gqlField */
-export function nodes(userConnection: PostConnection): Post[] {
+export function nodes(userConnection: Connection<Post>): Post[] {
   return userConnection.edges.map((edge) => edge.node);
 }
-
-/** @gqlType */
-type PostEdge = {
-  /** @gqlField */
-  node: Post;
-  /** @gqlField */
-  cursor: string;
-};
 
 // --- Root Fields ---
 
@@ -43,7 +27,7 @@ export async function posts(
     before?: string | null;
   },
   ctx: Ctx,
-): Promise<PostConnection> {
+): Promise<Connection<Post>> {
   const rows = await DB.selectPosts(ctx.vc);
   const posts = rows.map((row) => new Post(row));
   return connectionFromArray(posts, args);
