+++ examples/production-app/models/UserConnection.ts
@@ -3,31 +3,15 @@ import * as DB from "../Database";
 import { Ctx } from "../ViewerContext";
 import { Query } from "../graphql/Roots";
 import { User } from "./User";
-import { PageInfo } from "./PageInfo";
+import { Connection } from "../graphql/Connection";
 import { connectionFromArray } from "graphql-relay";
-
-/** @gqlType */
-export type UserConnection = {
-  /** @gqlField */
-  edges: UserEdge[];
-  /** @gqlField */
-  pageInfo: PageInfo;
-};
 
 /**
  * Convenience field to get the nodes from a connection.
  * @gqlField */
-export function nodes(userConnection: UserConnection): User[] {
+export function nodes(userConnection: Connection<User>): User[] {
   return userConnection.edges.map((edge) => edge.node);
 }
-
-/** @gqlType */
-type UserEdge = {
-  /** @gqlField */
-  node: User;
-  /** @gqlField */
-  cursor: string;
-};
 
 // --- Root Fields ---
 
@@ -43,7 +27,7 @@ export async function users(
     before?: string | null;
   },
   ctx: Ctx,
-): Promise<UserConnection> {
+): Promise<Connection<User>> {
   const rows = await DB.selectUsers(ctx.vc);
   const users = rows.map((row) => new User(row));
   return connectionFromArray(users, args);
